use wordnet
Could not figure out a way to use the installed library on it's own -
~wninit~ function is not properly defined in the ~libWN.a~. So the
advise from the documentation is actually pretty useless. Instead it
should read more like "copy the whole library, application is actually
the part of the library, you can't use the library on it's own, so
compile everything at once, good luck"

The command line interface is a good example of a simple application
that uses several WordNet library functions.