use different thread capping mechanism
Previous implementation only limited a number of the blame sections
running in parallel, but thread count wasn't limited in any way, which
still allowed for 40k+ threads to be created at once, effectively
halting any proper data processing. Now, cap is used on the ~async_task~
implementation - right now this is a kludge around ~condition_variable~,
but this really looks like a job for a semaphore.

Changed implementation of the logging/task launching solution - now
full information is passed to the task launcher in a single object,
completion logging is performed in there as well.