wip: user-provided python script for file fitering
Added command-line option handling and partial implementation for
user-provided filter script handing. More complex repositories are bound
to have a non-trivial rules about what parts of the project need to be
handled and how - ignored, processed etc. Trying to wrap all of this
complexity in globs/regexes would be pretty futile, and embedding python
interpreter in the C++ code is comparatively easy thanks to
boost::python library. This commit provides a partial implementation
for the embedding.